32-bit warning fixes in the debugger/win UI.

+++ Windows/main.cpp
@@ -101,7 +101,7 @@ int WINAPI WinMain(HINSTANCE _hInstance, HINSTANCE hPrevInstance, LPSTR szCmdLin
 	// Really should find a way to do this in XP too :/
 	if (0 != GetLocaleInfo(LOCALE_NAME_USER_DEFAULT, LOCALE_SNAME, lcCountry, 256)) {
 		langRegion = ConvertWStringToUTF8(lcCountry);
-		for (int i = 0; i < langRegion.size(); i++) {
+		for (size_t i = 0; i < langRegion.size(); i++) {
 			if (langRegion[i] == '-')
 				langRegion[i] = '_';
 		}
